Zaimplementowałem wzorzec Metoda Wytworcza w klasie Powiadomienie. Naprawiłem błąd w konstruktorze Glownego Kontrolera.
